refactor: move mouse tracker into util
